Cast popularity to float in all admin API queries
PostgreSQL's numeric(10,3) type is returned as a string by node-pg,
causing frontend .toFixed() calls to fail. Rather than adding
Number() wrappers on every frontend usage, fix at the source by
casting popularity::float in SQL queries.

This fixes the ActorManagementPage crash when filtering by popularity,
and preemptively fixes the same issue in:
- ActorDiagnosticPage
- HighPriorityActorsPage
- CacheManagementPage
- Various enrichment pages

+++ server/src/routes/admin/cache.ts
@@ -105,7 +105,7 @@ router.post("/warm", async (req: Request, res: Response): Promise<void> => {
     // Build query to get popular actors
     const deceasedFilter = deceasedOnly ? "WHERE a.deathday IS NOT NULL" : ""
     const query = `
-      SELECT a.id, a.name, a.deathday, a.popularity
+      SELECT a.id, a.name, a.deathday, a.popularity::float
       FROM actors a
       ${deceasedFilter}
       ORDER BY a.popularity DESC NULLS LAST
